requirements.txt
pyodbc==4.0.39
numpy==1.24.3
twilio==8.1.0
requests==2.31.0
python-dotenv==1.0.0
